Add read-only memory component
The read-only memory component has a variable address width and 4,8,16,
or 32 bit data width. The data is specified as a string of hexadecimal
characters. Whitespace is ignored.

The user-supplied hex characters are converted into an Array of values
when the component is created or edited. To decide the state of the
output values, the input values are used to computer an array index. The
index is used to look up the value in the array. Finally the value is
sent to the outputs.